changed map to support retina displays, and moved coordinate space conversion file to utilities
