+++ src/components/pixi-tephra-cross-section.tsx
@@ -8,7 +8,7 @@ import * as Color from "color";
 import * as L from "leaflet";
 import { getGridIndexForLocation } from "../stores/simulation-store";
 import gridTephraCalc from "../tephra2";
-import { LatLngToLocal, getDistanceFromLatLonInKm } from "./coordinateSpaceConversion";
+import { LatLngToLocal, getDistanceFromLatLonInKm } from "../utilities/coordinateSpaceConversion";
 
 interface IProps {
   canvasMetrics: ICanvasShape;
